Split bluetooth component and gamepad.
