+++ slider/src/bluetoothComponent.h
@@ -3,9 +3,7 @@
 
 #if !IS_SIMULATOR
 
-#include <ArduinoController.h>
-
-#include "input.h"
+#include "bluetoothGamepad.h"
 #include "core/component.h"
 #include "core/address.h"
 
@@ -13,26 +11,6 @@ class Controller;
 
 namespace Bt
 {
-    class BluetoothComponent;
-
-    class BluetoothGamepad
-    {
-    public:
-        IO::InputData ReadInput();
-        bool IsConnected() const;
-        bool HasData() const;
-
-        void SetPlayerLEDs(const uint8_t led);
-        void Rumble(const uint16_t delayedStartMs, const uint16_t durationMs, const uint8_t weakMagnitude,
-                    const uint8_t strongMagnitude);
-
-    private:
-        friend class BluetoothComponent;
-        BluetoothGamepad(Controller** controller);
-        IO::InputData m_LastInput = IO::InputData();
-        Controller** m_Controller;
-    };
-
     class BluetoothComponent : public Core::Component
     {
     public:
